refactor: add server folder
